refactor: Refatoração da estrutura de pacotes do projeto
